refactor: move factory.py to rag root directory
- Move src/crewai/rag/config/factory.py to src/crewai/rag/factory.py
- Update imports in utils.py and test_factory.py
